requirements.txt
streamlit
langchain
langgraph
langchain-google-genai
langchain-community
chromadb
pypdf2
python-dotenv
sentence-transformers
torch